Fix XML formatting from outside the editor (Source > Format). DO NOT MERGE.
This changeset fixes this issue:

20450: Source->Format for XML files in Project Explorer in 14-preview
       applies old XML rules instead of new Android-aware ones

Basically, the Eclipse XML support does not consult the
StructuredTextViewerConfigurationXML configuration when formatting
files.  To work around this, this changeset overrides the format
processor defined for XML files and redirects to the ADT custom
formatter when the option to use a custom formatter is set.

Unfortunately, setting the formatProcessor in the ADT plugin does not
work, because it turns out that the WTP code to process layer
registrations processes plugins in alphabetical order of plugin names
(rather than say plugin dependency order), and as a result, the WTP
registration is processed later than the ADT one (because org.eclipse
> com.android) and it clobbers our registration.

To work around this, this changeset adds a *new* plugin, an overlay
for ADT which is named "overlay.com.android.eclipse.(etc)" which sorts
alphabetically later than org.eclipse. The plugin is basically empty
(it just provides a registration of the ADT format processor), and it
is registered as just a part of the ADT plugin so it should not be
exposed to users directly.

Change-Id: I05c617ca8e879d2a46de9def9f97f3f7a79884ce