SDK: Fix bug that prevents repo XSD upgrades
This fixes the bug in the current code.
We're going to switch to a different upgrade mechanism,
rendering this code useless, but I'd still have this
checked in for the record: the issue is that the
temporary node is mixed up and the attributes use the
element name instead of their attribute names.

Change-Id: I3abcfd476494ee2ddb8b75789b346e86556d8d85