SDK: fix SdkManager class not releasing FileInputStream
In various parts of the code base, I see code that creates
a FileInputStrea and never closes it. Unfortunately on
Windows that means the file will stay locked for as long
as the java process is alive.

In this case the SdkManager wasn't closing to input streams
which makes the source.properties files locked, and thus
it makes it impossible to upgrade/delete the corresponding
packages.

There are more occurences of this elsewhere in the code.

Change-Id: I9f0a854b210411757eee8b4ebf8308b059206542