SDK: fix SdkManager class not releasing FileInputStream. Do not merge.
In various parts of the code base, I see code that creates
a FileInputStrea and never closes it. Unfortunately on
Windows that means the file will stay locked for as long
as the java process is alive.

In this case the SdkManager wasn't closing to input streams
which makes the source.properties files locked, and thus
it makes it impossible to upgrade/delete the corresponding
packages.

There are more occurences of this elsewhere in the code.

(cherry picked from commit 4c7bf6fe4dbf06f43d86b639ce5144ffeff7a626)

Change-Id: I36ad68a216b99fb471b941cb4ccdc88cd3bfad3b