SDK Manager: fix handling of source URLs
This changes how source URLs are handled. Packages
from different sources were previously treated as
separate if the packages were the same (e.g. same
platform API) but the source URLs were different.

Instead this checks the hostname+domain name of the URL
is different, as well as the type of the source (that
is a sdk repository vs addon repository).

Change-Id: I5ba133cd0dc984528bf696aef77d00cd9d31ed9c