SDK Manager: Prevent 'adb start-server' from blocking
SDK Manager hangs on windows after it tries to stop
and restart ADB (e.g. when installing platform-tools).

It hangs when capturing the stdout/stderr pipes -- these
don't close automatically when the process has finished
and these are not interruptible streams.
One workaround is to not capture them, especially since
the output isn't really useful for the installer anyway.

Change-Id: I6554461dfffad2cc8ff0f1fe7d212fdee742e2e6