Handle AAPT error messages for skipping files
If the AAPT error parser encounters an error message it is unfamiliar
with, then it aborts out of the parser and just emits all the output
from aapt in the Console instead.

I just ran into a scenario where this happens if your project contains
an XML file with an invalid name, such as a layout whose name starts
with an underscore.

This changeset adds a regexp to recognize these errors such that the
other AAPT errors show up in the Problems view, and you can click to
jump to the offending file for the skipping error.

Change-Id: I884fdb6fdbdfcaa3adbef68bc96e722036e88996