SDK Manager: don't output XML parse errors to stderr. Do not merge.
This is exactly like change I6f82b040 except now we
remove the stderr output when parsing the addons_list.xml

To put this in context, end-users should not see any
xml parsing error. In this case we get one because the
file doesn't exist on the test server and the web server
returns an HTML page describing the error, which naturally
fails to validate against our XSD. But since the point of
the method is just to validate, errors are irrelevant.

(cherry picked from commit 9d38047fb0c217cc3ee5ef01edf920ae642edb9a)

Change-Id: I31d86e9e7a61ce2fa595d5a2bcf4f8f76155decd