SDK Manager: don't output XML parse errors to stderr
This is exactly like change I6f82b040 except now we
remove the stderr output when parsing the addons_list.xml

To put this in context, end-users should not see any
xml parsing error. In this case we get one because the
file doesn't exist on the test server and the web server
returns an HTML page describing the error, which naturally
fails to validate against our XSD. But since the point of
the method is just to validate, errors are irrelevant.

Change-Id: I80b2c7600d1ceffa870ac3d71ef0c0f6ef9f809c