Use a scale widget to select a particular frame.
Currently, a slider (looks like a scrollbar) is present at the top
of the view, with no indication that it can be used to select the
frame to be displayed.

Replace that with the following layout:

Frame: -------------------------------- [23]

where the -'es are the scale widget, and the spinner at the end
indicates the currently selected frame. The spinner and the scale
are linked together to go up/down when either one changes.

Change-Id: Ie5403a179d518e4ebe50dba720c29c5bc07c4369