Ensure that custom views are unique
There are cases where Eclipse's code search engine returns multiple
instances of the same fully qualified class name when searching for
Views. In order not to show multiple versions of a custom view in the
palette, this changeset changes the CustomViewFinder from using a list
to using a set when accumulating class names.

Change-Id: I4dd9aafe76404efcf8b7bc82a5392f6010ba16f2