Improve indirect library dependency support.
Previously with the following setup:
    Main app depends on Lib1, Lib2.
    Lib1 and Lib2 both depends on Lib3.
Lib3 would be compiled 3 times, as the main app would attempt
to build it and then Libr1, and Lib2 would do it too.

Of course it wasn't actually built 3 times. After the first time it would
do nothing, checking dependencies and do nothing. However, for projects
will a lot of libraries and a lot of indirect dependencies this could
slow down a lot compilations, as the build would spend its time checking
whether to build the same library many times.

This ensure that only the main project attempts to build all of its library
dependencies (direct and indirect), in the right order of course.
When libraries are built they are told (through the "only" target) to not
build their dependencies.

Change-Id: Ie80f32a96e3c473d897aca5d05bf6fe9ea086884