Fix for Create Android Project from Existing Project
The project location was not updated when just pasting into the text
field; you had to use the file chooser.

In addition, copy out the inferred project name read from the manifest
(unless the user has already defined a project name), and don't
clobber a custom location if the user has manually edited the
location.

Change-Id: I4e70b7fcf49d2d2fb73d4da1f78104e92928b0e9