Lint improvements
First, in the translation detector, split out the report into two
separate issues:
- Failing to provide a translation for a string defined by the default
  locale (Error)
- Unused translations (strings found in languages but no longer in the
  default locale) (Warning)
Also add region handling such that a region can define just a subset
of the available strings and fall back to the overall language for
strings it does not define.

Second, the Quickfix in Eclipse for the px to dp conversion now shows
a dedicated dialog which has a combobox for the available densities
rather than asking the user to type in the density number in a text
field.

Third, the CLI driver (the "lint" command line tool) has a bunch of
new flags for controlling the report format as well as for dumping
out information about available issues:

Usage: lint [flags] <project directories>

Flags:
--suppress <id-list>: Suppress a list of issue id's.
--enable <id-list>: Only check the specific list of issues
--fullpath : Use full paths in the error output
--lines : Include the lines with errors in the output

--list: List the available issue id's and exit.
--show: List available issues along with full explanations
--show <ids>: Show full explanations for the given list of issue id's
Id lists should be comma separated with no spaces.

Here's an example of what an issue explanation looks like:

% lint --show DuplicateIds
DuplicateIds
------------
Summary: Checks for duplicate ids within a single layout or within an
include hierarchy

Priority: 7 / 10
Severity: Warning
Category: Layout

It's okay for two independent layouts to use the same ids. However,
within a single layout (including the case where two separate layouts
are fused together with an include tag) the ids should be unique such
that theActivity#findViewById() method can work predictably.

---

The layout editor now runs lint on every edit operation instead of
on every save.

And finally the scope has been moved from Detectors to Issues such
that when looking at issue markers we can clear those associated with
the scope the analysis is run under. We also run the detectors when
any of the associated issues is eligible.

Change-Id: I3f068de4edeb3ed166eeb7ab7f2cff84d7ae5cbd