New Lint Window
This changeset adds a new Lint View in Eclipse (available under Show
View in the Android category), which is a specialized view to view
lint warning markers.

This view opens automatically when you run Lint on a project. (The
view code is also shared with the single-file view shown when you
bring up the lint warnings from the layout editor.)

The lint view is similar to the Problems View (where lint markers also
show up), but with the following advantages:

(1) When you run lint, this view shows *only* lint warnings for the
    project, rather than being mixed in with Java warnings etc from
    this and other projects.

(2) The issue sorting is different here than in the Problems view: the
    results normally sort in order of priority, category, issue type,
    file and line number.

(3) The view contains a text area which displays more information for
    the currently selected issue, such as the detector summary and
    issue full explanation.

(4) The view toolbar has dedicated lint actions, such as a Fix
    lightbulb to run the associated fix (if any), an Ignore Type
    button for suppressing the selected type of issue, and Remove and
    Remove All buttons for clearing markers. And of course a "Refresh"
    button for re-running the analysis.

Change-Id: If58462d3eea198c07330ed9847f00925f346abaa