Two new lint warnings: array-size check, and manifest order
This changeset adds two warnings:

(1) A new lint warning which checks the sizes (number of items) in
    arrays of the same name across the various resource files and
    warns if the number of items differ.

(2) A new "manifest order" detector which looks for element ordering
    problems in manifest files, such as placing <uses-sdk> after the
    <application> element.

It also cleans up the way XML detectors were handled such that a
detector can be an XML detector without being a resource detector
(such as the manifest detector for example).

Finally it also moves classpath initialization into the tool context
where it belongs.

Change-Id: I02fee56bfdbe1064874acf70f18db79897d8a908