Add Lint HTML reporting, and fix position information in parser
This changeset adds a HTML report option for the lint command line
driver, and cleans up the code a bit. It also fixes the feature where
the line containing the error is shown along with each error, and
turns it on by default.

This was previously turned off because the position information for
XML nodes was fairly inaccurate (no column information, and the
element offset was the end of the opening tag rather than the
beginning).

With the HTML report, where not just the error line but the window of
code around the error is shown, having accurate offset information
even in the CLI driver was important. Therefore, this changeset also
rewrites the PositionXmlParser. The old approach used XML
transformers; now we instead use a SAX parser to incrementally build
up the DOM, and marking node offsets as we go along. SAX still only
gives element-end offsets, line+column instead of offsets, and no
information for individual attributes, so the builder also scans the
source in order to build up accurate information.

Results are now accumulated during the error reporting phase, then
sorted in priority, category and finally file+line number order before
being shown as HTML or text.

Finally, the code to escape code to attribute values was missing that
< is an invalid attribute character, so this CL fixes that issue as
well. Also fix some incorrect copyrights.

Change-Id: I652a1545b92865a88ecb4e6526616ba4cd75b512