Simple formatting fix
This changeset fixes #21383, #21346 and #21227 which are all related
to the fact that the formatter had assumed that the <item> tag will
only appear in resource files where it should appear on a single
line. However, <item> also appears in menu files and drawable-lists as
shown by these bugs.

This changeset just ensures that the <item>-special case is only
performed when using the resource file formatting style. It also adds
a regression test for this.

These testcases also reveal some other (lower priority) issues which
I'll fix in a follow-up CL.

Change-Id: I3194bbf4f556853eaf7a14c515ab3feb52177212