logcat: update scroll behavior when buffer is full.
When the logcat buffer is full, any received message will push out
an old message. As a result, the messages keep moving up the table,
which is annoying if you want to view a particular message.

Ideally, we'd keep the view constant, but only the scrollbar moves
up to indicate that we have received new messages. But this doesn't
work in an SWT table.

One solution would be to refresh the table and then scroll it to
remain in the same place. This causes a flicker and is annoying.

This patch stops refreshing the table when the user is viewing an
older item in the table, and starts auto refresh if the scroll bar
is again brought to the bottom. In addition it provides a toolbar
button that reflects the state of scrolling, similar to a scroll
lock checkbox.

The one downside to this is that since the scrollbar does not move,
the user does not have clue if new messages are being received.
The new messages will be noticed only if the user unselects the
scroll lock button, or manually moves the scrollbar down.

This is a fix for Issue #20805.

Change-Id: I12108789265fec38287a466d21505db85892594d