Fix refresh issues after a library build.
At the end of a library build we need to refresh the library
bin folder to ensure that the main project is notified of the changes.

However it looks like the refresh doesn't trigger a new build unless
it is asynchronous from a different job.

Also fixed the delta visitor used by the main project to look at
libraries to support looking into bin/res to detect output of the
crunch cache.

(Also added some debug output that is currently disabled).

Change-Id: Ide6d8700e3e98fca096051111b869b91fe2ac330