New lint icon detector
This changeset adds a new image detector which looks for various
issues related to icons:

(1) Are there bitmap icons in the densityless res/drawable/ folder?
(2) Do the icons appear in all the density folders (except,
    optionally, ldpi)
(3) Are the launcher/notification/etc icons of the expected size?
    (For example a launcher icon should be 48x48 in mdpi). This
    issue is disabled by default.
(4) Do the various bitmaps roughly have the same density-adjusted
    sizes across all the densities for the same configuration
    parameters? This will catch cases where assets are placed in the
    wrong place.
(5) It warns about using .gif files since that format is supported but
    discouraged.

Change-Id: Ibe0f97cba7cada5ac0deee244310a38dbc721873