Add lint check for accidental omission of android: prefix
This changeset adds a new detector which finds cases where you are
attempting to reference a view attribute in a layout on one of the
android views, but you forgot to add the android: prefix. For example,
you might attempt to set the "orientation" attribute on a
LinearLayout, which aapt will not complain about since it will
interpret it as a custom attribute on the target view.

There is also an Eclipse quickfix for this suggestion (as well as a
quickfix for the ScrollViewChildDetector).

This changeset also modifies the HTML report to bundle up associated
image resources in a local resource directory, and fixes a bug in the
category sort.

Change-Id: I984d1d575c0f9ec36fc450abf68ba75328cebf80