Add nodpi check
Adds a new icon check which makes sure that any icons which appear in
a -nodpi folder do not also appear in a density folder such as -mdpi.

This changeset also pulls the "missing density folder" out as a
separate issue from the "missing icon" issue where it warns about
icons that are present in some but not all density folders.

Change-Id: I0198613487100e2a9421b69eb55de690d8a814fd