Add gradual disclosure to the lint HTML report
When running lint on a large codebase the number of errors for any
category can be daunting, forcing you to scroll a lot to see what the
issues are. This changeset changes this such that for issues with more
than 8 warnings, only the first 5 are shown inline, and then there is
a "show more" button to disclose the rest.

Also fix a limitation in the array size detector which meant that it
didn't include a link to one of the files when detecting an array size
inconsistency.

Change-Id: I758879a0c6962f4663274e42a0c5a83ac1b8c8c6