Remove "unknown element" logging message
If you open the Error Log, and then go to (for example) a layout XML
file and type in a new tag, character by character, you will see that
the error log quickly fills up with messages of the form:

!ENTRY com.android.ide.eclipse.adt 2 0 2011-11-08 16:33:10.773
!MESSAGE AndroidManifest: Ignoring unknown 'Li' XML element

!ENTRY com.android.ide.eclipse.adt 2 0 2011-11-08 16:33:10.885
!MESSAGE AndroidManifest: Ignoring unknown 'Line' XML element

!ENTRY com.android.ide.eclipse.adt 2 0 2011-11-08 16:33:11.372
!MESSAGE AndroidManifest: Ignoring unknown 'Linear' XML element

What's happening is that on each keystroke, the UI model is updated,
and it complains if it sees element names that it does not recognize
from the descriptors. However, during an interactive edit it's normal
to have incomplete tag names, so we shouldn't complain -- aapt will do
that when the file is finally saved and processed.

Change-Id: Ib314ed0e1e17e69e35f05858ced5043fca2c29f3