Minor lint tweaks
First, replace tabs with spaces in the quoted error lines in the
command line output, since otherwise the column marker (^) which is
space based may not line up with the part of the line it is supposed
to point to.

Second, make the code which looks up the project corresponding to a
selection able to find a project when you click on something *in* a
project, such as a Java file within a Java project.

Third, don't print "skipping file because it contains parser errors"
in the CLI output; there's no need to since the parser themselves
report parsing errors now.

Change-Id: I57fd8a783c422a287cf9529caa1f858d7cff52d1