21750: Layout editor auto-format breaks rendering of custom views
The XML editor has a special "ignore model updates" flag which it uses
to skip doing work. This flag is set when we perform changes to the
editor that we know will have no semantic changes, such as
reformatting the XML.

Turns out there was a path where this flag did not get read properly,
which meant that when auto-formatting was enabled, it would perform
two renders. For reasons that are not clear, rendering custom views
does not work properly when the model is rendered a second time, but
since that render shouldn't happen anyway, the bug is fixed by
ensuring that we always skip model updates when we know they are only
cosmetic.

Change-Id: I49478e2dd7d68d63c4129ab1fe0b2ff5e9d208d1