Add two new layout detectors
This changeset adds two new layout_weight related detectors for LinearLayouts:

(1) Nested weights. This looks for cases where you have a weight on a
    child of a LinearLayout that is defined within an outer view which
    also has a weight. This is very inefficient because of the way
    views need to get measured when weights ar used.

(2) Baseline alignments. When you're using LinearLayouts as a way to
    proportionally distribute available space for children layouts
    (e.g. the children are not baseline views like buttons or text
    views) then it's more efficient to set baselineAligned=false on
    the LinearLayout. This also adds an Eclipse quickfix for this.

Also makes a shared fix class for the many various lint fixes that
just set an attribute. There will be more of these over time.

Change-Id: Ia6082d2436a156f8d63d6930af98a2ed363550e6