Add support for custom lint detectors
This changeset adds a mechanism where additional detectors can be
added by the user, either via an environment variable
(ANDROID_LINT_JARS) or by placing jar files in the ~/.android/lint/
directory. Each jar file is scanned for a manifest entry
(Lint-Registry) which points to an implementation of an IssueRegistry,
and the issues provided by that registry via getIssues() is added to
the global list of issues to be checked.

Change-Id: I0ac79a0d7dea50ae4258f70399b297afb761d4f8