Ensure that the conversion suggestion only appears for nonconst
We have a Java quick assistant which recognizes non-constant
expression errors, and suggests to convert the surrounding switch
statement to an if-else chain.

It turns out that this quick assistant can make its suggestion for not
just non-constant expressions but other errors involving R-id's as
well, because even though the correction assistant only returns
hasCorrections=true for problemId=IProblem.NonConstantExpression,
Eclipse proceeds to call getCorrections() on other types of errors
anyway. So work around that by filtering on the error message as well.

Change-Id: Iae24ebffd3123800e80338ab27da780f1cdae42c