Add lint support for fallback configurations
This changeset adds support for the command line lint tool to provide
"fallback configurations". Lint already looks at any lint.xml files
found in projects it scans, but if you are going to scan multiple
projects you need to have a lint.xml in each project.

This changeset adds a -config <file> option where you can point to a
lint configuration file which will be used as a fallback for all
projects, so if no specific severity is set for a particular issue in
a given project lint.xml file, then it will fall back to the setting
in the global lint.xml file and use that instead.

This makes it project to have a master configuration used across lots
of projects as part of a build step.

Change-Id: I3a8d13d9ffaff04b22649b6464f97faf84d265d5