Set android:ems="10" for EditTexts on drop
Text fields (without a text attribute) has by default only a single
character visible. To work around this, we automatically set the
layout_width to match_parent instead of wrap_content in LinearLayouts,
but in RelativeLayout and GridLayout etc (and in general) we don't
have this, so text fields show up with a not particularly usable size.

By setting the ems attribute to 10, the wrap_content size of the text
field will show 10 characters which seems like a reasonable default
(and when editing the XML or properties it's more obvious what you'd
want to modify to get a different size, if not match_parent.

Change-Id: Ic0460cea7810076b416f6d7d25a683851465ee9b