Add package support for Rules API and Detector API
When writing custom view rules, and custom detectors, you may need
access to the package name of the application (in order to produce the
right namespace to look up custom attributes for).

This changeset adds new methods to the two APIs such that clients can
look up the application package.  For lint, it also changes things
such that the manifest file is always processed and information
stashed in the Project object. This meant I could also remove the icon
detector code to look up the manifest info directly.

Change-Id: Id8eec96e2788f6d27481d7dcab0c544ea6e0a06d