SDK: Fix "APK missing" on "skip-package" build/launch.
Steps to reproduce:
- set ADT in skip-packaging build mode
- open a project, remove its bin dir, build once (fine), deploy.
- do Project > Clean... partial build happens.
- deploy: rest of complication happens and deploy fails with
  "Could not find blah.apk"
However the APK has been successfully created.

Solution:
Because the post compiler builder does a delayed refresh due to
library not picking the refresh up if it's done during the build,
we want to force a refresh here as this call is generally asking for
a build to use the apk right after the call.

Change-Id: If7ee7a1e1db5cb1802f9ce11836c91735784b997