21005: getContext() of custom view is null and throws NPE
Issue 21005 shows an exception thrown from the pull parser during
style processing. From the exception it looks like the pull parser's
getAttributeValue() method is called with invalid parameters.

However, it looks like the rest of the method codes defensively
against a null localName argument, so we might as well do it for the
last case as well.

Change-Id: Idb32e471ef7e14b5010ba5b85f900eb377a87a4f