21837: Layout Editor XML Formatting Creates Newlines
This changeset fixes
   21837: Layout Editor XML Formatting Creates Newlines When It
          Encounters Invalid Characters

The bug report talks about invalid characters in the XML, but when you
format a document with invalid characters the invalid characters are
converted to text. And the real bug here is that text nodes containing
newlines could result in additional newlines being inserted on every
formatting operation.

The fix is simple: strip prefix and suffix newlines from text nodes
before they are printed.

Change-Id: I4cc925288f56530a898dac34377cee2e7020a51a