Fix for custom view rendering with auto-format, revisited
This changeset revisits
http://code.google.com/p/android/issues/detail?id=21750
"Layout editor auto-formatting breaks rendering of custom views"

The bug described in the issue is that dragging custom views into the
canvas does not trigger a render until you save the file. My initial
bug fix noted that we actually would get *two* render requests, and
the second render was blank, so I fixed the model notifications such
that only a single notification (triggering a render) was sent.

However, collapsing the two changeModel calls into the DOM model
turned out to have some adverse effects when auto-format was enabled,
so I turned off auto-format until I could track the bug down.

In this changeset I'm turning it back on, and fixing the root cause
which made the custom view not appear: in the XML DOM to UI Model
sync, no descriptor was found for the custom view so the node simply
wasn't added to the model.

We still don't want multiple notifications (and therefore renders), so
I'm also changing the notifications structure a bit. Now a single
notification is made after all the edit hooks are run, and after the
edit lock has been released (it grabs a read lock instead for the
duration of the notification phase).

Change-Id: I2281dc936bf0442e2d38ac2750e81a8f120a924e