ADT Launch: default to always showing device chooser dialog.
The device chooser dialog has a checkbox that says use the same
device for future launches of an application. This is currently
enabled, which does not match current user expectation that the
dialog is always shown.

This patch disables the checkbox by default, so that the behavior
is identical to previous versions of ADT.

Change-Id: I32c9e903f0dbdb8c7ff1da38714125d58bfa06a9