Fix auto-indent for text nodes after closing tag (issue #22332)
This changeset fixes the auto-indent handling for a specific scenario:
pressing newline within a text node, when the next non-text region is
an element end tag, and the caret is not on the same line as the end
tag. This scenario is covered by the new unit tests testIssue22332a
and testIssue22332b.

Change-Id: I238423ddf25864746a941d6ac8c1fb6a41e37601