Add flags to lint cli to turn off warnings, and turn on all
Add gcc-style flags -w, -Wall and -Werror, and javac-style --nowarn
(identical to -w) to turn off warnings (only check for errors), enable
all warnings, and to turn all warnings into errors respectively.

Also cleans up the print-usage code a little bit such that the format
is nicer: the arguments all appear on the left and the explanations
are all aligned on the right.

Change-Id: I247f9e81ebe5322c7b19af2f043c7a82133c18a1