Make Fix in Lint Warnings window work for closed files
When you click on the Fix icon in the Lint Warnings window, the marker
you're applying a fix for may be referencing a file that is not open
and visible. The apply fix code needs to account for this, and open
the document (such that getModelForEdit will succeed) and for
interactive fixes, also open the editor.

Change-Id: Ic9c232bd050e4d6cce8acfc41a8e24b66e21bb48