SDK Manager: verbose print if unzip fails.
The core issue was that source packages were not properly
zipped on the build server side, which made the unzip crash
with an EOFException. This simply adds a better verbose
message in case this happens again.

It seems the Java ZipFile bug listed in there is fixed
at least in the JVM I'm currently running (1.6 and 7)
but anyway I added the extra safeguard code as documented
in the reported bug (e.g. don't ask ZipFile to unzip more
than the actual size of the entry.) Can't hurt to have it.

Change-Id: Ib47530920474e320da69c35d9a695d931bed55c1