SDK Windows: more ways to find java.
Also properly redirect command-line output
to the parent shell console.

This is experimental. This CL removes "android.exe"
from the Windows tools and only keeps android.bat.
However in a next CL android.bat will be changed
to use 'find_java' to locate the best Java exe around
(this is currently done in a bat script)

Change-Id: I6e5485fdf59fde9838cf929ff333e1c611ea7bb4