SDK Manager: 'list sdk' show system images and source pkg.
This fixes the 'android list sdk' to output the system
images and source packages in the list. Since there are
no hard dependencies on them, they were not listed.

The behavior is changed to:
- 'android list sdk': display all potential and
  mandatory updates and their dependencies.
- 'android list sdk --obsolete': display everything
  compatible on the remote sources.

This allows users to install sources and system images
from the command line:
$ android list sdk -e -o
$ android update sdk -u -o -t source,system-image # all
$ android update sdk -u -o -t source-14,sysimg-14 # API14

Also, although system images are not mandatory (e.g. they
are not hard dependencies of platforms), the heuristic is
that if a platform doesn't have any system image of its
own we'll suggest any matching sys-image we can find for
either new platform or the currently highest installed one.

SDK Bug: 21880

Change-Id: I4c47c6e60d885ebeaf181288db743b382fdb1618