gltrace: do not store images in memory
Protocol buffer messages can optionally contain the contents of the
framebuffer. These messages are then really large in size. Rather
than storing all of these messages in memory, we remove the image
data from the message and store a pointer (offset in file) to the
image in file. When needed, the trace file is read again to retrieve
the image.

Change-Id: I3d0c65026bba29efc5e31d8c251582fdaa068c64