rcp: create monitor plugin & build scripts
The monitor plugin defines the entry point for the monitor RCP application,
and also defines the monitor.product product configuration.

The build scripts in eclipse/scripts/rcp show how to build the RCP app
from the command line.

Currently, these are not built on the build server. That will happen once
we figure out where to put the prebuilts.

Change-Id: Idc96b2c011dba55c0d6792e916193017df6ce1c1