logcat: when copying items, index into filtered list.
The previous commit introduced a bug: The SWT table model is a list of
logcat messages before any filters are applied. But the user sees the
filtered messages. So when user selects items in the table, the
selection index will not match the index in the input model.

This patch applies all current filters to the input model and then
indexes this filtered list to get the same items that the user has
selected in the view.

Change-Id: If955bd5c22e1c587bcf5fe32092f513d53c0d8ef