New lint rule for invalid @+id references
This changeset adds a new layout rule which finds cases where
you have RelativeLayout attributes pointing to id names that do
not exist, such as
    layout_alignLeft="@+id/navbar"

The +id/ syntax here creates the id on the fly, so aapt does not
complain, but the new lint rule makes sure that at least one layout
actually defines this id as its id.

If no match is found, it looks for spelling mistakes and suggests
other similar ids in the error message.

This condition is flagged as an error. The detector also issues a
related warning if it finds an id reference that is not defined in the
same layout. This is sometimes okay (where you have deliberately
referred to an element which will be included into this layout, or is
in a layout including this one), but it's usually an accident.

Change-Id: I7f3fa6f4cb5cef6dcd307256050fb9580887addd