logcat: Improve auto scroll lock behavior
Setting up scroll lock automatically is broken due to the highly
platform dependent nature of the scroll bar events. This patch
improves the current situation significantly, but may not be the
last word on this subject.

This patch does the following:
1. Sets up auto scroll lock on Windows and Linux.
2. Removes auto scroll locking on Mac. Users have to manually
press the Pause Logcat button.
3. Renames the "Scroll to latest" button to "Pause Logcat" to
more accurately reflect what is happening.

Change-Id: I1ce03371d7ac7aeaea218aa3c5c426c69c56a527