Always run png crunch from Eclipse.
png crunch is done by aapt which does it only on the files that
need it (new or updated file). So it's safer to always call it
from eclispe rather than trying to figure out if it's needed
from the resource delta (which could be incomplete).
In particular, right now, the resource delta only checks for
the source folder.

Change-Id: I74448eee2f826346ca5a67e8894fafa414eba559