Fix Run Lint toolbar action to pick up more selection contexts
You currently have to select a project in the package explorer before
you click on the Run Lint toolbar action. This changeset makes the
action look harder for which project to run lint for:

- As before, look for the current selection and check whether it
  provides a selection
- If there's an active editor, use its project.
- If not, if there's just one available Android project, use it.
- If not, if there's just one non-library Android project, use it.

Finally, display a warning dialog if no project can be found, or if
the selected project is not an Android project, such that it's
apparent why nothing happens.

Change-Id: I92daa080db0bd815bc3d51fe9e458df12e6cc50f