Do not initialize DDMS Console from ADT plugin.
When ADT plugin starts up, it attempts to set the console to be used
for certain parts of DDMS to be the Android console.

Doing this however causes the DDMS plugin to be activated, which
launches adb server if it is not already there.

Removing this enables us to not touch adb unnecessarily.

This patch moves the initialization of the DDMS console into the
DDMS plugin, and uses the DDMS console as opposed to the Android
console for messages from DDMS.

Change-Id: I010a7028a2f22ac4da1c55903e001dcdd329d91f