Fix Lint false positive for mixed XML/PNG Bitmap resource
This changeset fixes this issue:
http://code.google.com/p/android/issues/detail?id=23214

The icon detector makes sure that if an icon appears in a given -dpi
folder, it appears in all the other -dpi folders as well (except for
-nodpi). However, in the set comparison it would use the full filename
(including file extension), which is not correct since it's okay to
have an .9.png in one folder and a .gif in another (and as reported in
this bug, an .xml file).

This changeset also associates a file location with folder warnings
for a missing density folder (it uses the res folder), which made
various unit tests need updates since the results sort differently.

Change-Id: I7f09772f3a54683a9d0c5c5d93ae2499707f533a