Remove JDT dependency from DDMS & TraceView.
Currently, there are three locations that require the ability
to open a Java file containing a particular method: DDMS stack trace,
logcat and traceview.

The logcat and traceview panels use JDT to open up a Java editor on
a method. As a result, these plugins have a dependency on JDT.

The stack trace panel uses the DDMS plugin's sourceRevealer extension
point to achieve the same purpose. ADT plugin provides an
implementation of the sourceRevealer plugin. As a result, this
introduces a JDT dependency only on the ADT plugin.

This patch makes both the logcat & trace view code to use the same
sourceRevealer extension point. As a result, all source code
involved in revealing the source of a Java method is present in
a single place, and DDMS & Traceview plugins don't depend on JDT
anymore.

Change-Id: I5b840dc4d6ad45e5611a80cff740101c484a99dd